refactor: add AuthServiceImpl for DIP
