+++ src/main/java/com/rudkids/rudkids/domain/auth/application/JwtTokenProvider.java
@@ -1,5 +1,6 @@
 package com.rudkids.rudkids.domain.auth.application;
 
+import com.rudkids.rudkids.domain.auth.application.TokenProvider;
 import com.rudkids.rudkids.domain.auth.exception.ExpiredTokenException;
 import com.rudkids.rudkids.domain.auth.exception.InvalidTokenException;
 import io.jsonwebtoken.*;
